refactoring of popup components => popup + popup-comments
